+++ local-cli/core/index.js
@@ -10,9 +10,15 @@
  */
 'use strict';
 
+const android = require('./android');
 const Config = require('../util/Config');
+const findPlugins = require('./findPlugins');
+const findAssets = require('./findAssets');
+const ios = require('./ios');
+const windows = require('./windows');
+const wrapCommands = require('./wrapCommands');
 
-const defaultConfig = require('./default.config');
+const flatten = require('lodash').flatten;
 const minimist = require('minimist');
 const path = require('path');
 
@@ -35,6 +41,61 @@ export type RNConfig = {
   getDependencyConfig(pkgName: string): Object,
 };
 
+const getRNPMConfig = (folder) =>
+  // $FlowFixMe non-literal require
+  require(path.join(folder, './package.json')).rnpm || {};
+
+const attachPackage = (command, pkg) => Array.isArray(command)
+  ? command.map(cmd => attachPackage(cmd, pkg))
+  : { ...command, pkg };
+
+const defaultRNConfig = {
+  getProjectCommands(): Array<CommandT> {
+    const appRoot = process.cwd();
+    const plugins = findPlugins([appRoot])
+      .map(pathToCommands => {
+        const name = pathToCommands.split(path.sep)[0];
+
+        return attachPackage(
+          // $FlowFixMe non-literal require
+          require(path.join(appRoot, 'node_modules', pathToCommands)),
+          // $FlowFixMe non-literal require
+          require(path.join(appRoot, 'node_modules', name, 'package.json'))
+        );
+      });
+
+    return flatten(plugins);
+  },
+
+  getProjectConfig(): Object {
+    const folder = process.cwd();
+    const rnpm = getRNPMConfig(folder);
+
+    return Object.assign({}, rnpm, {
+      ios: ios.projectConfig(folder, rnpm.ios || {}),
+      android: android.projectConfig(folder, rnpm.android || {}),
+      windows: windows.projectConfig(folder, rnpm.windows || {}),
+      assets: findAssets(folder, rnpm.assets),
+    });
+  },
+
+  getDependencyConfig(packageName: string) {
+    const folder = path.join(process.cwd(), 'node_modules', packageName);
+    const rnpm = getRNPMConfig(
+      path.join(process.cwd(), 'node_modules', packageName)
+    );
+
+    return Object.assign({}, rnpm, {
+      ios: ios.dependencyConfig(folder, rnpm.ios || {}),
+      android: android.dependencyConfig(folder, rnpm.android || {}),
+      windows: windows.dependencyConfig(folder, rnpm.windows || {}),
+      assets: findAssets(folder, rnpm.assets),
+      commands: wrapCommands(rnpm.commands),
+      params: rnpm.params || [],
+    });
+  },
+};
+
 /**
  * Loads the CLI configuration
  */
@@ -44,7 +105,7 @@ function getCliConfig(): RNConfig {
     ? Config.loadFile(path.resolve(__dirname, cliArgs.config))
     : Config.findOptional(__dirname);
 
-  return {...defaultConfig, ...config};
+  return {...defaultRNConfig, ...config};
 }
 
 module.exports = getCliConfig();
